+++ Sources/arm/ui/TabFonts.hx
@@ -133,7 +133,9 @@ class TabFonts {
 				ui._y += 6;
 			}
 
-			if (ui.isDeleteDown && Project.fonts.length > 1 && Context.font.file != "") {
+			var inFocus = ui.inputX > ui._windowX && ui.inputX < ui._windowX + ui._windowW &&
+						  ui.inputY > ui._windowY && ui.inputY < ui._windowY + ui._windowH;
+			if (inFocus && ui.isDeleteDown && Project.fonts.length > 1 && Context.font.file != "") {
 				ui.isDeleteDown = false;
 				deleteFont(Context.font);
 			}
