+++ Sources/arm/ui/TabLayers.hx
@@ -373,7 +373,9 @@ class TabLayers {
 				}
 			}
 
-			if (ui.isDeleteDown && canDelete(Context.layer)) {
+			var inFocus = ui.inputX > ui._windowX && ui.inputX < ui._windowX + ui._windowW &&
+						  ui.inputY > ui._windowY && ui.inputY < ui._windowY + ui._windowH;
+			if (inFocus && ui.isDeleteDown && canDelete(Context.layer)) {
 				ui.isDeleteDown = false;
 				deleteLayer(Context.layer);
 			}
